Consolidate all SAML class code to single file
